@@ -59,7 +59,7 @@ SciRate::Application.configure do
   # config.action_controller.asset_host = "http://assets.example.com"
 
   # Precompile additional assets (application.js, application.css, and all non-JS/CSS are already added)
-  config.assets.precompile << /\.(?:svg|eot|woff|ttf)$/
+  # config.assets.precompile << /\.(?:svg|eot|woff|ttf)$/
 
   # Ignore bad email addresses and do not raise email delivery errors.
   # Set this to true and configure the email server for immediate delivery to raise delivery errors.
