@@ -17,14 +17,14 @@ ActiveRecord::Schema.define(version: 20170523032959) do
   enable_extension "plpgsql"
 
   create_table "auth_links", force: :cascade do |t|
-    t.string   "provider",         limit: 255, null: false
-    t.string   "uid",              limit: 255, null: false
-    t.string   "oauth_token",      limit: 255, null: false
-    t.datetime "oauth_expires_at",             null: false
-    t.integer  "user_id",                      null: false
+    t.string   "provider",         null: false
+    t.string   "uid",              null: false
+    t.string   "oauth_token",      null: false
+    t.datetime "oauth_expires_at", null: false
+    t.integer  "user_id",          null: false
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.text     "auth",                         null: false
+    t.text     "auth",             null: false
   end
 
   create_table "authors", force: :cascade do |t|
@@ -110,15 +110,15 @@ ActiveRecord::Schema.define(version: 20170523032959) do
   add_index "comments", ["user_id"], name: "index_comments_on_user_id", using: :btree
 
   create_table "delayed_jobs", force: :cascade do |t|
-    t.integer  "priority",               default: 0, null: false
-    t.integer  "attempts",               default: 0, null: false
-    t.text     "handler",                            null: false
+    t.integer  "priority",   default: 0, null: false
+    t.integer  "attempts",   default: 0, null: false
+    t.text     "handler",                null: false
     t.text     "last_error"
     t.datetime "run_at"
     t.datetime "locked_at"
     t.datetime "failed_at"
-    t.string   "locked_by",  limit: 255
-    t.string   "queue",      limit: 255
+    t.string   "locked_by"
+    t.string   "queue"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
@@ -196,7 +196,7 @@ ActiveRecord::Schema.define(version: 20170523032959) do
   add_index "scites", ["user_id"], name: "index_scites_on_user_id", using: :btree
 
   create_table "sessions", force: :cascade do |t|
-    t.string   "session_id", limit: 255, null: false
+    t.string   "session_id", null: false
     t.text     "data"
     t.datetime "created_at"
     t.datetime "updated_at"
@@ -220,8 +220,8 @@ ActiveRecord::Schema.define(version: 20170523032959) do
     t.text     "alert",                default: "",                    null: false
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.datetime "arxiv_sync_dt",        default: '2015-04-12 00:00:00', null: false
-    t.datetime "arxiv_author_sync_dt", default: '2015-04-12 00:00:00', null: false
+    t.datetime "arxiv_sync_dt",        default: '2020-12-25 00:00:00', null: false
+    t.datetime "arxiv_author_sync_dt", default: '2020-12-25 00:00:00', null: false
   end
 
   create_table "users", force: :cascade do |t|
