@@ -21,8 +21,10 @@ describe CommentsController do
     before do
       become user
       expect(Net::HTTP).to receive(:start).once.and_return(double(code: '200'))
-      xhr :post, :create, comment: { paper_uid: paper.uid, content: "fishies" }
-      expect(response).to be_redirect
+
+      post :create, params: { comment: { paper_uid: paper.uid, content: "fishies" } }, xhr: true
+
+      expect(response).to redirect_to(paper)
     end
 
     it "creates a comment" do
@@ -36,11 +38,11 @@ describe CommentsController do
   describe "editing a comment" do
     before do
       become user
-      xhr :post, :edit, id: comment.id, content: "wubbles"
+      post :edit, params: { id: comment.id, content: "wubbles" }, xhr: true
     end
 
     it "edits the comment" do
-      expect(response).to be_success
+      expect(response).to be_successful
       expect(comment.reload.content).to eq "wubbles"
     end
   end
@@ -48,8 +50,8 @@ describe CommentsController do
   describe "deleting a comment" do
     before do
       become user
-      xhr :post, :delete, id: comment.id
-      expect(response).to be_redirect
+      post :delete, params: { id: comment.id }, xhr: true
+      expect(response).to redirect_to(paper)
     end
 
     it "marks comment as deleted" do
@@ -64,8 +66,8 @@ describe CommentsController do
       become user
       comment.deleted = true
       comment.save
-      xhr :post, :restore, id: comment.id
-      expect(response).to be_redirect
+      post :restore, params: { id: comment.id }, xhr: true
+      expect(response).to redirect_to(paper)
     end
 
     it "restores the comment" do
@@ -77,8 +79,8 @@ describe CommentsController do
   describe "replying to a comment" do
     before do
       become user
-      xhr :post, :reply, id: comment.id, content: "snuffles"
-      expect(response).to be_redirect
+      post :reply, params: { id: comment.id, content: "snuffles" }, xhr: true
+      expect(response).to redirect_to(paper)
     end
 
     it "creates a new comment in reply" do
@@ -94,19 +96,19 @@ describe CommentsController do
 
     it "allows a single upvote" do
       expect do
-        xhr :post, :upvote, id: comment.id
-        expect(response).to be_success
-        xhr :post, :upvote, id: comment.id
+        post :upvote, params: { id: comment.id }, xhr: true
+        expect(response).to be_successful
+        post :upvote, params: { id: comment.id }, xhr: true
         comment.reload
       end.to change(comment, :cached_votes_up).by(1)
     end
 
     it "allows unvoting" do
       expect do
-        xhr :post, :upvote, id: comment.id
-        expect(response).to be_success
-        xhr :post, :unvote, id: comment.id
-        expect(response).to be_success
+        post :upvote, params: { id: comment.id }, xhr: true
+        expect(response).to be_successful
+        post :unvote, params: { id: comment.id }, xhr: true
+        expect(response).to be_successful
         comment.reload
       end.to_not change(comment, :cached_votes_up)
     end
