@@ -42,7 +42,7 @@ describe User do
   it { is_expected.to validate_presence_of(:fullname) }
   it { is_expected.to validate_length_of(:fullname).is_at_most(50) }
   it { is_expected.to validate_presence_of(:email) }
-  it { is_expected.to validate_uniqueness_of(:email) }
+  it { is_expected.to validate_uniqueness_of(:email).case_insensitive }
 
   describe "when email format is invalid" do
     invalid_addresses = %w[user@foo,com user_at_foo.org example.user@foo.]
